added documentation for part_1, added makefile content for part_1, finished homework part_1

+++ exercise_2/src/part_1/run_and_plot.py
@@ -1,34 +1,43 @@
 # -*- coding: utf-8 -*-
 import time
 import os
+import numpy as np
 import matplotlib as mp
 
-av = []
-ts = []
+av = np.zeros(23)
+ts = np.zeros(23)
 n = range(2,25)
+readonly = False
+numberMeasurements = 10.0
 
-if not os.access("./main.c",os.F_OK):
-    raise EnvironmentError("main.c is missing!")
+if not readonly:
+    os.system("scp ./main.c hpc_creek08:~/chr.c")
+    os.system("ssh hpc_creek08 'mpicc -Wall ~/chr.c -o ~/chr'")
+    for i in range(int(numberMeasurements)):
 
-os.system("scp ./main.c hpc_creek08:~/chr.c")
-os.system("ssh hpc_creek08 'mpicc -Wall ~/chr.c -o ~/chr'")
+        for i,nn in list(enumerate(n)):
+            print "starting executable with",nn,"processes"
+            pipe = os.popen("ssh hpc_creek08 'mpirun -host creek01,creek02,creek03,creek04,creek05,creek06,creek07,creek08 -np "+str(nn)+" ~/chr'",'r')
+            output = pipe.read()
+            print "the output for",nn,"processes is:", output
+            pipe.close()
+            totalTime = float(output.split()[0])
+            averageTime = float(output.split()[1])
+            ts[i] += totalTime
+            av[i] += 1000*averageTime
 
-for nn in n:
-    print "starting executable with",nn,"processes"
-    pipe = os.popen("ssh hpc_creek08 'mpirun -host creek01,creek02,creek03,creek04,creek05,creek06,creek07,creek08 -np "+str(nn)+" ~/chr'",'r')
-    time.sleep(2)
-    output = pipe.read()
-    print "the output for",nn,"processes is:", output
-    pipe.close()
-    totalTime = float(output.split()[0])
-    averageTime = float(output.split()[1])
-    ts.append(totalTime)
-    av.append(1000*averageTime)
+    os.system("ssh hpc_creek08 'rm ~/chr'")
+    os.system("ssh hpc_creek08 'rm ~/chr.c'")
+    ts /= numberMeasurements
+    av /= numberMeasurements
 
-os.system("ssh hpc_creek08 'rm ~/chr'")
-os.system("ssh hpc_creek08 'rm ~/chr.c'")
-
-
+    if os.access("./plotdata.dat",os.F_OK):
+        os.remove("./plotdata.dat")
+    f = open("./plotdata.dat",'w')    
+    for nn,avav,tsts in zip(n,av,ts):
+        f.write(str(nn)+" "+"{:.10f}".format(avav)+" "+"{:.10f}".format(tsts)+"\n")
+    print "Wrote plotdata to file ..."
+    f.close()    
 
 mp.use("pgf")
 pdf_with_custom_preamble = {
@@ -44,21 +53,38 @@ pdf_with_custom_preamble = {
         ]
 }
 
+f = open("./plotdata.dat",'r')
+content = f.read()
+n = []
+av = []
+ts = []
+for line in content.split("\n")[:-1]:
+    splittedline = line.split()
+    nn = float(splittedline[0])
+    avav = float(splittedline[1])
+    tsts = float(splittedline[2])
+
+    n.append(nn) 
+    av.append(avav)
+    ts.append(tsts)
+
+print "read plotdata from file"
+f.close()    
 
 mp.rcParams.update(pdf_with_custom_preamble)
 
 import matplotlib.pyplot as plt
 
-fig = plt.figure(figsize=(4,8))
+fig = plt.figure(figsize=(5,6))
 
 ax1 = fig.add_subplot(211)
 ax2 = fig.add_subplot(212)
 ax1.plot(n,ts)
 ax1.set_xlabel(r"number of processes $n$")
-ax1.set_ylabel("total communication\ntime [s]")
+ax1.set_ylabel(r"total communication time [s]")
 ax2.plot(n,av)
 ax2.set_xlabel(r"number of processes $n$")
-ax2.set_ylabel("average time per\nmessage [ms]")
+ax2.set_ylabel(r"average time per message [ms]")
 fig.subplots_adjust(left = 0.2, hspace=0.0)
 ax1.xaxis.grid()
 ax2.xaxis.grid()
